feat: Add click tap listener to lightweight charts for mobile
